Whops, messed up damage value types. v 0.02

+++ src/com/github/toxuin/PVPItem.java
@@ -4,7 +4,7 @@ import org.bukkit.inventory.ItemStack;
 
 public class PVPItem {
 	public Integer id;
-	public Integer damage;
+	public double damage;
 	public String name;
 	public ItemStack getItemStack () {
 		return new ItemStack(this.id);
